ml-service/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
python-docx
spacy
sentence-transformers
scikit-learn
numpy
pandas
pymongo
motor
python-jose
passlib
bcrypt
python-dotenv
httpx